sfos: Chunk PBAP pull requests
Do not supply all PBAP content to a client's pull request in one
response, but request data from obexd-calldata-provider in chunks and
pass them along to the client one by one. This brings down the latency
between request from a client and the first response packet, even if
the total time to serve the request may slightly increase due to
increased roundtrips.